# AuthAPI/src/main/resources/swagger/auth-api.yaml
openapi: 3.0.3
info:
  title: API d'Authentification
  description: Service d'authentification JWT pour l'architecture microservices
  version: 1.0.0
  contact:
    name: EL OUAZIZI Walid
    email: admin@walidlab.dev

servers:
  - url: http://localhost:8080/api/v1
    description: Serveur de développement

tags:
  - name: Authentification
    description: Gestion de l'authentification et des sessions
  - name: Validation
    description: Validation des tokens JWT

paths:
  /signup:
    post:
      tags:
        - Authentification
      summary: Créer un nouveau compte utilisateur
      description: Enregistre un nouvel utilisateur dans le système
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: L'utilisateur existe déjà
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /signin:
    post:
      tags:
        - Authentification
      summary: Se connecter au système
      description: Authentifie un utilisateur et retourne un token JWT
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /logout:
    delete:
      tags:
        - Authentification
      summary: Se déconnecter du système
      description: Révoque le token JWT en l'ajoutant à la liste noire
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Déconnexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          description: Token invalide ou manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /check:
    get:
      tags:
        - Validation
      summary: Vérifier la validité d'un token
      description: Valide un token JWT et retourne son statut
      operationId: checkToken
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token valide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResponse'
        '401':
          description: Token invalide, expiré ou révoqué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT au format "Bearer {token}"

  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Nom d'utilisateur unique
          example: waldo_dev
        password:
          type: string
          minLength: 8
          format: password
          description: Mot de passe (minimum 8 caractères)
          example: SecureP@ss123
        email:
          type: string
          format: email
          description: Adresse email valide
          example: waldo@example.com
        firstName:
          type: string
          description: Prénom
          example: Walid
        lastName:
          type: string
          description: Nom
          example: EL OUAZIZI

    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indique si l'inscription a réussi
          example: true
        message:
          type: string
          description: Message de confirmation
          example: Utilisateur créé avec succès
        user:
          $ref: '#/components/schemas/User'

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Nom d'utilisateur
          example: waldo_dev
        password:
          type: string
          format: password
          description: Mot de passe de l'utilisateur
          example: SecureP@ss123

    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indique si la connexion a réussi
          example: true
        token:
          type: string
          description: Token JWT généré
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          description: Type de token
          example: Bearer
          default: Bearer
        user:
          $ref: '#/components/schemas/User'

    LogoutResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indique si la déconnexion a réussi
          example: true
        message:
          type: string
          description: Message de confirmation
          example: Déconnexion réussie, token révoqué

    CheckResponse:
      type: object
      properties:
        valid:
          type: boolean
          description: Indique si le token est valide
          example: true
        username:
          type: string
          description: Nom d'utilisateur associé au token
          example: waldo_dev

    User:
      type: object
      properties:
        username:
          type: string
          description: Nom d'utilisateur
          example: waldo_dev
        email:
          type: string
          format: email
          description: Adresse email
          example: waldo@example.com
        firstName:
          type: string
          description: Prénom
          example: Walid
        lastName:
          type: string
          description: Nom
          example: EL OUAZIZI

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Code d'erreur
          example: AUTH_001
        message:
          type: string
          description: Message d'erreur descriptif
          example: Identifiants incorrects
        timestamp:
          type: string
          format: date-time
          description: Horodatage de l'erreur
          example: 2024-12-13T14:30:00Z